Move unroll_place to Place::unroll
